feat: optimize raster performance on Animated images. (#513)
The I/O bottleneck can impact the performance of the raster thread in
Flutter.

When WebF pages contain multiple GIF images, each frame of these GIFs
requires the I/O thread to fetch its next frame, potentially turning the
I/O thread into a bottleneck.

This PR introduces an intersectionObserverLayer for each GIF image,
monitoring the visibility of each animated image. If an ImageElement
scrolls outside the viewport, the imageFrame event listeners will be
detached from the imageStreamListener, and the next frame request for
that image will be paused.

With this update, only the animated images visible in the viewport will
request the next frame, significantly reducing the load on the I/O
thread.

Related issue for Flutter:
https://github.com/flutter/flutter/issues/135443